API 0.12.2 Se modificó método grade_relCAL de rel_score_service.py para corregir bugs de error 422 en componente front de relCAL_Calificaciones.jsx (Revisar commit FRONT 0.9.0 para más información) La modificación que hizo método anterior mencionado fue quitar estudFK de los requerimientos de consulta y petición, dejando solo a actividFK Se modificó rel_score_model, cambiando la nota a tipo float en correspondencia con la BD

+++ models/rel_score_model.py
@@ -4,7 +4,7 @@ from pydantic import BaseModel
 class Rel_Score(BaseModel):
     estudFK: int
     actividFK: int
-    nota: int
+    nota: float
     archivo_url: str
     comentario: str
     feedback: str
